sample-app/requirements.txt
flask==2.3.3
requests==2.31.0
opentelemetry-api==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-exporter-otlp==1.19.0
opentelemetry-instrumentation-flask==0.40b0
opentelemetry-instrumentation-requests==0.40b0
opentelemetry-instrumentation-logging==0.40b0
prometheus-client==0.17.1
pyroscope-io==0.8.7